pyproject.toml
[tool.poetry]
name = "datawarehouse-jd"
version = "0.1.0"
description = ""
authors = ["Victor Zucchetti <victorzucchetti2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pre-commit = "^3.7.1"
pandas = "^2.2.2"
sqlalchemy = "^2.0.31"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
yfinance = "^0.2.40"
dbt-postgres = "^1.8.2"
dbt-core = "^1.8.3"
streamlit = "^1.36.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
